# cloud/output: Atlas friendly benchmarks