@codebien 2022-09-16
The increment of `NewSampleFromTrail` sounds fine because we added the following new lines in the atlas branch:
https://github.com/grafana/k6/blob/be8ca6fe61c5961d83399a76f49d5aee07e94a91/output/cloud/data.go#L114-L118